+++ Example/PixelTest/Views/SimpleView/SimpleView.swift
@@ -26,14 +26,3 @@ final class SimpleView: UIView {
     }
     
 }
-
-extension UIView {
-    
-    static var nib: UINib {
-        return UINib(nibName: "\(classForCoder())", bundle: Bundle(for: self))
-    }
-    
-    static func loadFromNib<T>() -> T {
-        return nib.instantiate(withOwner: nil, options: nil).first as! T
-    }
-}
